added urls support

+++ yk/config.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 import os
+import random
 import tomllib
 from functools import reduce
 from pathlib import Path
@@ -12,41 +13,42 @@ from validators import url as is_url
 from .util import YTA_Q, con, pf
 
 
-def parse_configs(files: list = [], cfg_to_del: dict = {}, args=None):
-    tomls = []
+def parse(i: list = [], args=None, cfg_to_del: dict = {}):
+    o = []
 
-    for file in files:
+    for file in i:
         toml = {}
 
         #########################
         ## toml validation
 
         try:
-            with open(file, 'rb') as f:
-                toml = tomllib.load(f)
+            if Path(file).is_file():
+                with open(file, 'rb') as f:
+                    toml = tomllib.load(f)
+                    if not toml:
+                        continue
 
-            assert toml  # empty check
+            elif is_url(file):
+                toml = {file: {}}
 
         except tomllib.TOMLDecodeError as ex:
-            log.error(f'toml decode error in {file!r}, {ex}')
+            log.exception(f'toml decode error in {file!r}, {ex}')
             continue
 
         except FileNotFoundError as ex:
-            log.error(f"file {file!r} not found', {ex}")
+            log.exception(f"file {file!r} not found', {ex}")
             continue
 
         except AssertionError:
-            log.error(f'file {file!r} is empty')
+            log.exception(f'file {file!r} is empty')
             continue
 
         except Exception as ex:
-            log.error(f'file error in {file!r}, {ex}')
+            log.exception(f'file error in {file!r}, {ex}')
             continue
 
         # log.trace(f'stock {file}:\n' + pf(toml))
-
-        #########################
-        ## finding single-use items via '@' prefix in key
 
         for item in toml.copy().keys():
             if toml[item].get('url') or toml[item].get('u'):
@@ -72,36 +74,49 @@ def parse_configs(files: list = [], cfg_to_del: dict = {}, args=None):
         for item, cfg in toml.copy().items():
             # basic options
             toml[item]['url'] = cfg.get('url') or cfg.get('u') or ''
-            toml[item]['folder'] = cfg.get('folder') or cfg.get('f') or ''
             toml[item]['quality'] = cfg.get('quality') or cfg.get('q') or 'best'
+            toml[item]['output'] = cfg.get('output') or cfg.get('o') or args.output
+            toml[item]['folder'] = cfg.get('folder') or cfg.get('f') or ''
             toml[item]['delete'] = bool(cfg.get('delete') or cfg.get('d') or False)
-            toml[item]['health'] = bool(cfg.get('health') or False)
+            toml[item]['health'] = bool(cfg.get('health') or cfg.get('h') or False)
 
-            # checking method
+            rgx = cfg.get('regex') or cfg.get('r') or ''
+            toml[item]['regex_title'] = cfg.get('regex_title') or rgx
+            toml[item]['regex_desc'] = cfg.get('regex_desc') or rgx
+
+            toml[item]['proxy'] = cfg.get('proxy') or (
+                random.choice(args.proxy) if args.proxy else ''
+            )
+            toml[item]['apprise'] = cfg.get('apprise') or args.apprise
+            toml[item]['cookies'] = cfg.get('cookies') or args.cookies
+            toml[item]['bgutil'] = cfg.get('bgutil') or args.bgutil
+
+            # live-stream checking method
             toml[item]['checker'] = cfg.get('checker') or cfg.get('chk') or args.chk
             if toml[item]['checker'] not in ['str', 'dlp']:
                 toml[item]['checker'] = args.chk
 
             # recording method
             toml[item]['recorder'] = cfg.get('recorder') or cfg.get('rec') or args.rec
-
             if toml[item]['recorder'] not in ['str', 'dlp', 'yta']:
                 toml[item]['recorder'] = args.rec
 
+            # check for ytarchive recorder for non-youtube streams
             if toml[item]['recorder'] == 'yta' and not con(
                 ['youtube.com', 'youtu.be'], toml[item]['url']
             ):
                 log.warning(
-                    f'recording method is ytarchive, but non-youtube url detected ({toml[item]["url"]}), will fallback to yt-dlp',
+                    f'{file}: {item}: recording method is ytarchive, but non-youtube url detected ({toml[item]["url"]}), will fallback to yt-dlp',
                     item=toml[item],
                 )
                 toml[item]['recorder'] = 'dlp'
 
-            # arguments for recording process
+            # cli arguments for recorder
             toml[item]['arguments'] = (
                 cfg.get('record_arguments') or cfg.get('rec_args') or None
             )
 
+            # choosing specific cli args for recorder
             if not toml[item]['arguments']:
                 match toml[item]['recorder']:
                     case 'str':
@@ -113,36 +128,43 @@ def parse_configs(files: list = [], cfg_to_del: dict = {}, args=None):
                     case 'yta':
                         toml[item]['arguments'] = args.yta_args
 
+            # 'worst' not supported in ytarchive
             if (
                 toml[item]['recorder'] == 'yta'
                 and toml[item]['quality'] not in YTA_Q
                 and ('youtube' in toml[item]['url'] or 'youtu.be' in toml[item]['url'])
             ):
                 log.error(
-                    f"{toml[item]['quality']!r} not supported in ytarchive, fallback to 'best' for now",
+                    f"{file}: {item}: {toml[item]['quality']!r} not supported in ytarchive, fallback to 'best' for now",
                     file=file,
                     item=item,
                 )
                 log.info(f'choose something from {", ".join(YTA_Q)!r}')
                 toml[item]['quality'] = 'best'
 
-            rgx = cfg.get('regex') or cfg.get('r') or ''
-            toml[item]['regex_title'] = cfg.get('regex_title', rgx)
-            toml[item]['regex_desc'] = cfg.get('regex_desc', rgx)
-
             if not toml[item]['url']:
-                log.trace(f'{file}:{item}: empty url, skipping.')
+                # log.trace(f'{file}:{item}: empty url, skipping.')
                 toml.pop(item)
                 continue
 
             if not is_url(toml[item]['url']):
                 log.warning(
-                    f'{file}:{item}: {cfg["url"]!r} is not a valid url, skipping.'
+                    f'{file}: {item}: {cfg["url"]!r} is not a valid url, skipping.',
                 )
                 toml.pop(item)
                 continue
 
-            for i in ('u', 'f', 'r', 'q', 'd'):
+            try:
+                out_path = Path(toml[item]['output'])
+                out_path.mkdir(parents=True, exist_ok=True)
+                toml[item]['output'] = str(out_path.resolve())  # subprocess pwd fix
+            except:  # noqa: E722
+                log.exception(
+                    f'{file}:{item}: failed to create output dir: {toml[item]["output"]}'
+                )
+                continue
+
+            for i in ('u', 'f', 'r', 'q', 'd', 'h', 'o'):
                 if i in cfg:
                     del toml[item][i]
 
@@ -158,6 +180,9 @@ def parse_configs(files: list = [], cfg_to_del: dict = {}, args=None):
                     toml_or.pop(item)
 
                     path = Path(file)
+                    if not path.is_file():
+                        continue
+
                     fst = path.stat()
 
                     with open(file, 'wb') as f:
@@ -170,7 +195,10 @@ def parse_configs(files: list = [], cfg_to_del: dict = {}, args=None):
 
         log.trace(f'{file}:\n' + pf(toml))
 
-        tomls.append(toml)
+        o.append(toml)
+
+    if not o:
+        return o
 
     # return merged dicts
-    return reduce(lambda x, y: x | y, tomls)
+    return reduce(lambda x, y: x | y, o)
